Support Debian/Ubuntu for core_begin cube

+++ core_begin/core_begin.sh
@@ -48,28 +48,30 @@ echo ""
 grep -e MemTotal -e MemFree -e Buffers -e ^Cached /proc/meminfo
 echo ""
 
-cube_package install firewalld
-
-cube_service start firewalld
-
-cube_service enable firewalld
-
 cube_include firewall_whitelist
 
 # Machines may be memory constrained, so disable crons for the duration
 # of the chef-client run. Re-enable in the core_end cube
-cube_check_dir_exists "/etc/cron.d" && cube_service stop crond
+if [ "$(systemctl is-active crond)" = "active" ]; then
+  cube_service stop crond
+fi
+
+if cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_DEBIAN} ; then
+  cube_package update
+  cube_package install selinux-utils
+fi
 
 # Ensure permissive selinux
-if cube_check_dir_exists "/etc/selinux/" && cube_set_file_contents "/etc/selinux/config" "templates/selinux_config" ; then
+if cube_check_dir_exists "/etc/selinux/" && cube_set_file_contents "/etc/selinux/config" "templates/selinux_config" && [ "$(getenforce)" != "Disabled" ] ; then
   setenforce Permissive || cube_check_return
 fi
 
-cube_set_file_contents_string ~/.passwd ${cubevar_app_passwords_root}
-
 # We set a user password in the case we need to do a manual login from the web console
-passwd --stdin root < ~/.passwd
-rm -f ~/.passwd
+echo "root:${cubevar_app_passwords_root}" | chpasswd
+
+#cube_set_file_contents_string ~/.passwd ${cubevar_app_passwords_root}
+#passwd --stdin root < ~/.passwd || cube_check_return
+#rm -f ~/.passwd
 
 if cube_set_file_contents "/etc/commonprofile.sh" "templates/commonprofile.sh" ; then
   chmod 755 "/etc/commonprofile.sh" || cube_check_return
@@ -86,26 +88,54 @@ if cube_set_file_contents "/etc/systemd/journald.conf" "templates/journald.conf"
   cube_service restart systemd-journald
 fi
 
-cube_package install python python-dnf multitail htop lsof wget nfs-utils at
+if cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_FEDORA}; then
+  cube_package install python python-dnf multitail htop lsof wget nfs-utils at
+elif cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_DEBIAN}; then
+  cube_package install python multitail htop lsof wget nfs-common at apt-transport-https
+else
+  cube_throw Not implemented
+fi
 
 cube_service enable atd
 cube_service start atd
 
-cube_package --enablerepo fedora-debuginfo --enablerepo updates-debuginfo install kernel-debuginfo-common-x86_64 kernel-debuginfo glibc-debuginfo-common glibc-debuginfo systemtap perf
+if cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_FEDORA}; then
+  cube_package --enablerepo fedora-debuginfo --enablerepo updates-debuginfo install kernel-debuginfo-common-x86_64 kernel-debuginfo glibc-debuginfo-common glibc-debuginfo systemtap perf
+elif cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_DEBIAN}; then
+  # https://wiki.ubuntu.com/DebuggingProgramCrash#Debug_Symbol_Packages
+  if ! cube_check_file_exists /etc/apt/sources.list.d/ddebs.list ; then
+    echo "deb http://ddebs.ubuntu.com $(lsb_release -cs) main restricted universe multiverse" | tee -a /etc/apt/sources.list.d/ddebs.list
+    # The following is in one of the Ubuntu guides but returns:
+    # W: The repository 'http://ddebs.ubuntu.com xenial-security Release' does not have a Release file.
+    #echo "deb http://ddebs.ubuntu.com $(lsb_release -cs)-security main restricted universe multiverse" | tee -a /etc/apt/sources.list.d/ddebs.list
+    echo "deb http://ddebs.ubuntu.com $(lsb_release -cs)-updates main restricted universe multiverse" | tee -a /etc/apt/sources.list.d/ddebs.list
+    echo "deb http://ddebs.ubuntu.com $(lsb_release -cs)-proposed main restricted universe multiverse" | tee -a /etc/apt/sources.list.d/ddebs.list
+    apt-key adv --keyserver keyserver.ubuntu.com --recv-keys 5FDFF622 || cube_check_return
+    apt-key adv --keyserver keyserver.ubuntu.com --recv-keys ECDCAD72428D7C01 || cube_check_return
+    apt-get update || cube_check_return
+  fi
 
-if cube_set_file_contents "/var/chef/cache/cookbooks/dnf/libraries/dnf-query.py" "templates/dnf-query.py" ; then
-  chmod 755 "/var/chef/cache/cookbooks/dnf/libraries/dnf-query.py" || cube_check_return
+  # https://wiki.ubuntu.com/Kernel/Systemtap
+  cube_package install linux-image-$(uname -r)-dbgsym systemtap linux-tools-generic
+else
+  cube_throw Not implemented
 fi
 
 # https://www.elastic.co/guide/en/logstash/current/installing-logstash.html
-cube_read_heredoc <<'HEREDOC'; cubevar_app_str="${cube_read_heredoc_result}"
+# https://docs.influxdata.com/telegraf/v1.1/introduction/installation/
+if cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_FEDORA}; then
+  cube_read_heredoc <<'HEREDOC'; cubevar_app_str="${cube_read_heredoc_result}"
 [influxdb]
 name = InfluxDB Repository - RHEL \$releasever
 baseurl = https://repos.influxdata.com/rhel/7Server/\$basearch/stable
 enabled = 1
 gpgcheck = 0
 gpgkey = https://repos.influxdata.com/influxdb.key
+HEREDOC
 
+  cube_set_file_contents_string "/etc/yum.repos.d/influxdb.repo" "${cubevar_app_str}"
+
+  cube_read_heredoc <<'HEREDOC'; cubevar_app_str="${cube_read_heredoc_result}"
 [logstash-5.x]
 name=Elastic repository for 5.x packages
 baseurl=https://artifacts.elastic.co/packages/5.x/yum
@@ -116,14 +146,52 @@ autorefresh=1
 type=rpm-md
 HEREDOC
 
-cube_set_file_contents_string "/etc/yum.repos.d/influxdb.repo" "${cubevar_app_str}"
+  cube_set_file_contents_string "/etc/yum.repos.d/logstash.repo" "${cubevar_app_str}"
+elif cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_DEBIAN}; then
+  if ! cube_check_file_exists /etc/apt/sources.list.d/elastic-5.x.list ; then
+    cube_app_tmp="$(wget -qO - https://artifacts.elastic.co/GPG-KEY-elasticsearch || cube_check_return)" || cube_check_return
+    printf '%s' "${cube_app_tmp}" | apt-key add - || cube_check_return
+    echo "deb https://artifacts.elastic.co/packages/5.x/apt stable main" | tee -a /etc/apt/sources.list.d/elastic-5.x.list || cube_check_return
+    cube_package update
+  fi
+  
+  if ! cube_check_file_exists /etc/apt/sources.list.d/influxdb.list ; then
+    cube_app_tmp="$(curl -sL https://repos.influxdata.com/influxdb.key || cube_check_return)" || cube_check_return
+    printf '%s' "${cube_app_tmp}" | apt-key add - || cube_check_return
+    . /etc/lsb-release || cube_check_return
+    echo "deb https://repos.influxdata.com/${DISTRIB_ID,,} ${DISTRIB_CODENAME} stable" | tee /etc/apt/sources.list.d/influxdb.list || cube_check_return
+    cube_package update
+  fi
+else
+  cube_throw Not implemented
+fi
 
-cube_package install multitail strace htop mtr traceroute patch atop sysstat \
-                     iotop gdb bind-utils ntp python sendmail make mailx \
-                     postfix tcpdump cyrus-sasl-plain rsyslog gnupg \
-                     kexec-tools lzo lzo-devel lzo-minilzo bison bison-devel \
-                     ncurses ncurses-devel telegraf telnet iftop git \
-                     nmap-ncat java-1.8.0-openjdk grub2-tools
+if cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_FEDORA}; then
+  cube_package install multitail strace htop mtr traceroute patch atop sysstat \
+                        iotop gdb bind-utils ntp python sendmail make mailx \
+                        postfix tcpdump cyrus-sasl-plain rsyslog gnupg \
+                        kexec-tools lzo lzo-devel lzo-minilzo bison bison-devel \
+                        ncurses ncurses-devel telegraf telnet iftop git \
+                        nmap-ncat java-1.8.0-openjdk grub2-tools
+elif cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_DEBIAN}; then
+  # https://wiki.ubuntu.com/Kernel/CrashdumpRecipe
+  # https://help.ubuntu.com/lts/serverguide/kernel-crash-dump.html
+
+  cube_read_heredoc <<'HEREDOC'; cubevar_app_kdump_tools="${cube_read_heredoc_result}"
+USE_KDUMP=1
+HEREDOC
+
+  cube_set_file_contents_string /etc/default/kdump-tools "${cubevar_app_kdump_tools}"
+  
+  cube_package install multitail strace htop mtr traceroute patch atop sysstat \
+                        iotop gdb ldap-utils ntp python make mailutils \
+                        postfix tcpdump rsyslog gnupg makedumpfile libsasl2-modules \
+                        kexec-tools liblzo2-2 liblzo2-dev libbison-dev \
+                        libncurses-dev telegraf telnet iftop git linux-crashdump \
+                        netcat-openbsd default-jdk `uname -r`-dbg crash
+else
+  cube_throw Not implemented
+fi
 
 cube_service enable atop
 cube_service start atop
@@ -133,23 +201,32 @@ if cube_set_file_contents "/etc/rsyslog.conf" "templates/rsyslog.conf" ; then
   cubevar_api_post_restart=$(cube_append_str "${cubevar_api_post_restart}" "rsyslog")
 fi
 
-cube_set_file_contents "/etc/kdump.conf" "templates/kdump.conf"
+if cube_service_exists kdump ; then
+  cube_set_file_contents "/etc/kdump.conf" "templates/kdump.conf"
 
-# don't auto-start because we may not have crashkernel yet
-cube_service enable kdump
+  # Don't auto-start because we may not have crashkernel yet
+  cube_service enable kdump
+fi
 
 # https://access.redhat.com/documentation/en-US/Red_Hat_Enterprise_Linux/7/html-single/Kernel_Crash_Dump_Guide/index.html#sect-kdump-memory-requirements
 # 160 MB + 2 bits for every 4 KB of RAM. For a system with 1 TB of memory, 224 MB is the minimum (160 + 64 MB).
 cubevar_app_crashkernel_mem=$((161+($(cube_total_memory)/268435456)))
 
-# https://docs.fedoraproject.org/en-US/Fedora/25/html/System_Administrators_Guide/sec-Making_Persistent_Changes_to_a_GRUB_2_Menu_Using_the_grubby_Tool.html
-# https://docs.fedoraproject.org/en-US/Fedora/25/html/System_Administrators_Guide/sec-Customizing_the_GRUB_2_Configuration_File.html
-#grubby --update-kernel=ALL "--args=no_timer_check console=hvc0 LANG=en_US.UTF-8 crashkernel=${cubevar_app_crashkernel_mem}M audit=0" || cube_check_return
-
-if [ -r /boot/extlinux/extlinux.conf ]; then
-  sed -i "s/UTF-8\$/UTF-8 crashkernel=${cubevar_app_crashkernel_mem}M audit=0/g" /boot/extlinux/extlinux.conf
+if cube_check_file_exists /boot/extlinux/extlinux.conf ; then
+  sed -i "s/UTF-8\$/UTF-8 crashkernel=${cubevar_app_crashkernel_mem}M audit=0/g" /boot/extlinux/extlinux.conf || cube_check_return
+elif cube_check_file_exists /boot/grub/grub.cfg ; then
+  if cube_set_file_contents "/etc/default/grub" "templates/grub1.template" ; then
+    update-grub || cube_check_return
+  fi
+  if cube_set_file_contents "/etc/default/grub.d/kexec-tools.cfg" "templates/kexec-tools.cfg.template" ; then
+    update-grub || cube_check_return
+  fi
 else
-  if cube_set_file_contents "/etc/default/grub" "templates/grub.template" ; then
+  # https://docs.fedoraproject.org/en-US/Fedora/25/html/System_Administrators_Guide/sec-Making_Persistent_Changes_to_a_GRUB_2_Menu_Using_the_grubby_Tool.html
+  # https://docs.fedoraproject.org/en-US/Fedora/25/html/System_Administrators_Guide/sec-Customizing_the_GRUB_2_Configuration_File.html
+  # grubby --update-kernel=ALL "--args=no_timer_check console=hvc0 LANG=en_US.UTF-8 crashkernel=${cubevar_app_crashkernel_mem}M audit=0" || cube_check_return
+
+  if cube_set_file_contents "/etc/default/grub" "templates/grub2.template" ; then
     /usr/sbin/grub2-mkconfig -o /boot/grub2/grub.cfg || cube_check_return
   fi
 fi
@@ -157,15 +234,36 @@ fi
 cube_ensure_directory ~/.ssh/ 700
 cube_ensure_file ~/.ssh/authorized_keys 700
 
-cube_package update
-cube_package --enablerepo fedora-debuginfo --enablerepo updates-debuginfo update
+if cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_FEDORA}; then
+  cube_package update
+  cube_package --enablerepo fedora-debuginfo --enablerepo updates-debuginfo update
+elif cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_DEBIAN}; then
+  cube_package update
+  cube_package upgrade
+else
+  cube_throw Not implemented
+fi
 
-cube_service enable ntpd
-cube_service start ntpd
+if cube_service_exists ntpd ; then
+  cube_service enable ntpd
+  cube_service start ntpd
+else
+  cube_service enable ntp
+  cube_service start ntp
+fi
+
+cubevar_app_hostname=$(cube_hostname)
 
 # To test sending mail from the box:
 #   echo "Message body" | mail -s "Subject" -r from@example.com to@example.com
-cube_set_file_contents "/etc/postfix/main.cf" "templates/main.cf.template"
+if cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_FEDORA}; then
+  cube_set_file_contents "/etc/postfix/main.cf" "templates/main.cf.fedora.template"
+elif cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_DEBIAN}; then
+  cube_set_file_contents_string "/etc/mailname" "myplaceonline.com"
+  cube_set_file_contents "/etc/postfix/main.cf" "templates/main.cf.debian.template"
+else
+  cube_throw Not implemented
+fi
 
 cube_service enable postfix
 cube_service start postfix
@@ -174,15 +272,17 @@ cube_set_file_contents "/etc/security/limits.conf" "templates/limits.conf"
 
 cube_set_file_contents ~/.toprc "templates/toprc"
 
-if ! cube_check_file_exists /usr/local/src/crash/crash ; then
-  cube_pushd /usr/local/src/
-  rm -rf crash* || cube_check_return
-  git clone https://github.com/crash-utility/crash/ || cube_check_return
-  cd crash || cube_check_return
-  echo '-DLZO' > CFLAGS.extra || cube_check_return
-  echo '-llzo2' > LDFLAGS.extra || cube_check_return
-  make || cube_check_return
-  cube_popd
+if ! cube_operating_system_has_flavor ${POSIXCUBE_OS_FLAVOR_DEBIAN} ; then
+  if ! cube_check_file_exists /usr/local/src/crash/crash ; then
+    cube_pushd /usr/local/src/
+    rm -rf crash* || cube_check_return
+    git clone https://github.com/crash-utility/crash/ || cube_check_return
+    cd crash || cube_check_return
+    echo '-DLZO' > CFLAGS.extra || cube_check_return
+    echo '-llzo2' > LDFLAGS.extra || cube_check_return
+    make || cube_check_return
+    cube_popd
+  fi
 fi
 
 if ! cube_has_role "syslog_server" ; then
